refactor(stable_api): extract HistoryMixin into pyquotex._api.history
Moves 9 history/candle methods (get_candles, _fetch_historical_batch,
_parse_historical_candles, get_historical_candles, get_candles_deep,
get_candle_v2, get_history_line, get_trader_history, prepare_candles)
from stable_api.py into a focused mixin. Public surface unchanged.
